fix: Fix compilation errors in risk assessment handler tests
- Removed non-existent errorHandler field from createTestHandler()
- Fixed Limit() call to use int instead of uint64 (postgrest expects int)
- All tests now compile and run successfully

+++ services/risk-assessment-service/internal/handlers/risk_assessment.go
@@ -871,7 +871,7 @@ func (h *RiskAssessmentHandler) HandleRiskTrends(w http.ResponseWriter, r *http.
 
 	// Limit results
 	if limit > 0 {
-		query = query.Limit(uint64(limit), "")
+		query = query.Limit(limit, "")
 	}
 
 	_, err := query.ExecuteTo(&assessments)
